refactor(spindle-tokens): change directory
